+++ api/app/signals/settings/base.py
@@ -68,7 +68,7 @@ INSTALLED_APPS = [
     'signals.apps.search',
     'signals.apps.dataset',
     'signals.apps.questionnaires',
-    'signals.apps.xperimental',
+    'signals.apps.experimental',
 
     'change_log',
